chore: split up ContactScreen, so that all related components and their tests lie together, and the 'dumb' screen components are together. #19
